+++ fabric-resource-loader-v0-bridge/src/main/java/net/fabricmc/fabric/api/resource/IdentifiableResourceReloadListener.java
@@ -1,0 +1,14 @@
+package net.fabricmc.fabric.api.resource;
+
+import java.util.Collection;
+import java.util.Collections;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.packs.resources.PreparableReloadListener;
+
+public interface IdentifiableResourceReloadListener extends PreparableReloadListener {
+    ResourceLocation getFabricId();
+
+    default Collection<ResourceLocation> getFabricDependencies() {
+        return Collections.emptyList();
+    }
+}