+++ fabric-resource-loader-v0-bridge/src/main/java/net/fabricmc/fabric/api/resource/ModResourcePack.java
@@ -1,0 +1,10 @@
+package net.fabricmc.fabric.api.resource;
+
+import net.fabricmc.loader.api.metadata.ModMetadata;
+import net.minecraft.server.packs.PackResources;
+
+public interface ModResourcePack extends PackResources {
+    ModMetadata getFabricModMetadata();
+
+    ModResourcePack createOverlay(String overlay);
+}