+++ fabric-resource-loader-v0-bridge/src/main/java/net/fabricmc/fabric/api/resource/ResourcePackActivationType.java
@@ -1,0 +1,11 @@
+package net.fabricmc.fabric.api.resource;
+
+public enum ResourcePackActivationType {
+    NORMAL,
+    DEFAULT_ENABLED,
+    ALWAYS_ENABLED;
+
+    public boolean isEnabledByDefault() {
+        return this == DEFAULT_ENABLED || this == ALWAYS_ENABLED;
+    }
+}