+++ fabric-resource-loader-v0-bridge/src/main/java/net/fabricmc/fabric/api/resource/SimpleSynchronousResourceReloadListener.java
@@ -1,0 +1,7 @@
+package net.fabricmc.fabric.api.resource;
+
+import net.minecraft.server.packs.resources.ResourceManagerReloadListener;
+
+public interface SimpleSynchronousResourceReloadListener
+        extends IdentifiableResourceReloadListener, ResourceManagerReloadListener {
+}